+++ mobile/android/android-components/docs/changelog.md
@@ -29,6 +29,9 @@ permalink: /changelog/
    * Added `CustomViewFinder`, a `View` that shows a ViewFinder positioned in center of the camera view and draws an Overlay
    * Added optional String resource `scanMessage` param to `QrFeature` for adding a message below the viewfinder
 
+* **service-experiments**
+  * ⚠️ **This is a breaking change**: Mako's configuration now requires explicitly setting an http client. Users need to pass one at construction.
+
 # 41.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v40.0.0...v41.0.0)